feat: RequiredAuth component to HOC with hook
