+++ packages/client/src/components/Layout/Layout.tsx
@@ -1,17 +1,12 @@
 import React from 'react'
 import { Outlet } from 'react-router-dom'
-import RequiredAuth from '@/components/RequiredAuth/RequiredAuth'
 
 interface ILayout {
   children?: JSX.Element
 }
 export const Layout = ({ children }: ILayout) => (
   <div className="layout">
-    <RequiredAuth>
-      <>
-        {children}
-        <Outlet />
-      </>
-    </RequiredAuth>
+    {children}
+    <Outlet />
   </div>
 )
